+++ CHANGELOG.md
@@ -12,7 +12,8 @@ Web UI version numbers should always match the corresponding version of LBRY App
   * New flow for rewards authentication failure
 
 ### Changed
-  * Reward-eligible content icon is now a rocket ship :D
+  * Make it clearer how to skip identity verification and add link to FAQ
+  * Reward-eligible content icon is now a rocket ship :D :D :D
   * Improved flow for when app is run with incompatible daemon
   *